requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
django-environ
django-filter
djangorestframework
drf-nested-routers
idna
Markdown
pillow
psycopg2-binary
requests
sqlparse
sslcommerz-lib
tzdata
urllib3
whitenoise